8.Gün ödevi

+++ Entity/Conctere/Car.cs
@@ -11,9 +11,9 @@ namespace Entity.Conctere
         public int CategoryId { get; set; } //Arabanın türü: spor,cip,sedan,hybrit
         public int ColorId { get; set; }
         public int BrandId { get; set; }
-        public long ModelYear { get; set; }
+        public int ModelYear { get; set; }
         public string Description { get; set; }
-        public double DailyPrice { get; set; }
+        public int DailyPrice  { get; set; }
 
     }
 }
